refactor: factor out git to services
