+++ apps/main-web-app/src/lib/services/git/index.ts
@@ -1,0 +1,3 @@
+export * from './clone';
+export * from './commit';
+export * from './push';
